read from excel and populate data in db
We injected an IWebHostEnvironment instance through dependency injection, just like we did for ApplicationDbContext, so that we can retrieve the web application path and be able to read the Excel file.
We added an Import action method that will use ApplicationDbContext and the EPPlus package to read the Excel file and add Countries and Cities.
At the start of the Import method’s implementation, we used the IWebHostEnvironment instance to determine if we’re running in a development environment or not: if we aren’t, the code will throw a SecurityException. By acting that way we’ll prevent anyone—including our users—from calling this method in production, thus restricting the whole importing task to developers only.
Countries are imported first because the City entities require the CountryId foreign key value, which will be returned when the corresponding Country is created in the database as a new record.
We defined a Dictionary container object to store all existing countries (plus each new Country right after we create it) so that we can query that list using LINQ to retrieve the CountryId instead of performing a lot of SELECT queries: this logic will also prevent the method from inserting the same country multiple times, should we happen to execute it more than once.
We defined another Dictionary container object to prevent the insertion of duplicate cities as well.
Last but not least, we created a JSON object to show the overall results on the screen.